[Edit Menu] 23 - 24 23. Undo: This is a one-level undo. It should be capable of undoing any operations that modify the data. e.g. Cut, Paste, Delete, Modulate and Rotate. The Undo menu should be grayed out if there is nothing to Undo. This can happen under 2 conditions: The document has not been modified, so there is nothing to Undo. The document has been modified more than once followed by one Undo. Since the Undo is a one-level operation, it means that once you Undo the last operation, all prior operations cannot be further undone.
24. Cut: Allows cutting the entire wave into the clipboard using a custom clipboard format. This allows applications that receive that custom format to paste it into them. At this point, the only application that would understand this format would be another instance of the same application. This should be grayed out if there is no data in the currently active document.
c. Copy: Allows copying into the clipboard of the entire wave using a custom clipboard format. This allows applications that receive that custom format to paste it into them. This should be grayed out if there is no data in the currently active document.
d. Copy As Bitmap: Allows copying the entire wave into the clipboard using the Bitmap format. This allows applications like Word/Visio to paste the bitmap into their documents.
e. Paste: Allows pasting of any wave data (Your previous custom clipboard wave format) present on the clipboard in place of the current wave document. This menu item should be grayed out if no Wave data is present on the clipboard.
f. Delete: Allows clearing of the entire wave. This item should be disabled if the document is empty.