[View Menu] 25-26; [6.Status bar] a-c
25. Toggle menus for viewing/hiding Status strip and Tool strip.
26. A Toggle menu item for viewing the wave either in standard mode or Full Mode.
a. Standard Mode: means each sample point maps to a single pixel. In all examples given the number of samples exceed the window size, so a scrollbar should be provided to scroll through the data.
b. Full/Zoomed mode, which allows viewing of the entire wave within the window. No scrollbars should be present.

6. The Status bar should have 4 panels:
a. A DropDownList-style Combo Box. This should hold any errors that occur during the application’s lifetime.
b. The number of currently open files.
c. The number of samples for the currently active MDI child window.